refactor(native): Abstract session property provider
