+++ presto-native-execution/presto_cpp/main/properties/session/SessionProperties.cpp
@@ -11,34 +11,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include "presto_cpp/main/SessionProperties.h"
+#include "presto_cpp/main/properties/session/SessionProperties.h"
+#include "presto_cpp/main/common/Utils.h"
 #include "velox/core/QueryConfig.h"
 
 using namespace facebook::velox;
 
 namespace facebook::presto {
 
-namespace {
-const std::string boolToString(bool value) {
-  return value ? "true" : "false";
-}
-} // namespace
-
 SessionProperties* SessionProperties::instance() {
   static std::unique_ptr<SessionProperties> instance =
       std::make_unique<SessionProperties>();
   return instance.get();
-}
-
-void SessionProperties::addSessionProperty(
-    const std::string& name,
-    const std::string& description,
-    const TypePtr& type,
-    bool isHidden,
-    const std::optional<std::string> veloxConfig,
-    const std::string& defaultValue) {
-  sessionProperties_[name] = std::make_shared<SessionProperty>(
-      name, description, type->toString(), isHidden, veloxConfig, defaultValue);
 }
 
 // List of native session properties is kept as the source of truth here.
@@ -53,7 +37,7 @@ SessionProperties::SessionProperties() {
       BOOLEAN(),
       false,
       QueryConfig::kExprEvalSimplified,
-      boolToString(c.exprEvalSimplified()));
+      util::boolToLowerCaseString(c.exprEvalSimplified()));
 
   addSessionProperty(
       kExprMaxArraySizeInReduce,
@@ -169,7 +153,7 @@ SessionProperties::SessionProperties() {
       BOOLEAN(),
       false,
       QueryConfig::kJoinSpillEnabled,
-      boolToString(c.joinSpillEnabled()));
+      util::boolToLowerCaseString(c.joinSpillEnabled()));
 
   addSessionProperty(
       kWindowSpillEnabled,
@@ -177,7 +161,7 @@ SessionProperties::SessionProperties() {
       BOOLEAN(),
       false,
       QueryConfig::kWindowSpillEnabled,
-      boolToString(c.windowSpillEnabled()));
+      util::boolToLowerCaseString(c.windowSpillEnabled()));
 
   addSessionProperty(
       kWriterSpillEnabled,
@@ -185,7 +169,7 @@ SessionProperties::SessionProperties() {
       BOOLEAN(),
       false,
       QueryConfig::kWriterSpillEnabled,
-      boolToString(c.writerSpillEnabled()));
+      util::boolToLowerCaseString(c.writerSpillEnabled()));
 
   addSessionProperty(
       kWriterFlushThresholdBytes,
@@ -203,7 +187,7 @@ SessionProperties::SessionProperties() {
       BOOLEAN(),
       false,
       QueryConfig::kRowNumberSpillEnabled,
-      boolToString(c.rowNumberSpillEnabled()));
+      util::boolToLowerCaseString(c.rowNumberSpillEnabled()));
 
   addSessionProperty(
       kMarkDistinctSpillEnabled,
@@ -211,7 +195,7 @@ SessionProperties::SessionProperties() {
       BOOLEAN(),
       false,
       QueryConfig::kMarkDistinctSpillEnabled,
-      boolToString(c.markDistinctSpillEnabled()));
+      util::boolToLowerCaseString(c.markDistinctSpillEnabled()));
 
   addSessionProperty(
       kSpillerNumPartitionBits,
@@ -228,7 +212,7 @@ SessionProperties::SessionProperties() {
       BOOLEAN(),
       false,
       QueryConfig::kTopNRowNumberSpillEnabled,
-      boolToString(c.topNRowNumberSpillEnabled()));
+      util::boolToLowerCaseString(c.topNRowNumberSpillEnabled()));
 
   addSessionProperty(
       kValidateOutputFromOperators,
@@ -240,7 +224,7 @@ SessionProperties::SessionProperties() {
       BOOLEAN(),
       false,
       QueryConfig::kValidateOutputFromOperators,
-      boolToString(c.validateOutputFromOperators()));
+      util::boolToLowerCaseString(c.validateOutputFromOperators()));
 
   addSessionProperty(
       kDebugDisableExpressionWithPeeling,
@@ -249,7 +233,7 @@ SessionProperties::SessionProperties() {
       BOOLEAN(),
       false,
       QueryConfig::kDebugDisableExpressionWithPeeling,
-      boolToString(c.debugDisableExpressionsWithPeeling()));
+      util::boolToLowerCaseString(c.debugDisableExpressionsWithPeeling()));
 
   addSessionProperty(
       kDebugDisableCommonSubExpressions,
@@ -259,7 +243,7 @@ SessionProperties::SessionProperties() {
       BOOLEAN(),
       false,
       QueryConfig::kDebugDisableCommonSubExpressions,
-      boolToString(c.debugDisableCommonSubExpressions()));
+      util::boolToLowerCaseString(c.debugDisableCommonSubExpressions()));
 
   addSessionProperty(
       kDebugDisableExpressionWithMemoization,
@@ -270,7 +254,7 @@ SessionProperties::SessionProperties() {
       BOOLEAN(),
       false,
       QueryConfig::kDebugDisableExpressionWithMemoization,
-      boolToString(c.debugDisableExpressionsWithMemoization()));
+      util::boolToLowerCaseString(c.debugDisableExpressionsWithMemoization()));
 
   addSessionProperty(
       kDebugDisableExpressionWithLazyInputs,
@@ -280,7 +264,7 @@ SessionProperties::SessionProperties() {
       BOOLEAN(),
       false,
       QueryConfig::kDebugDisableExpressionWithLazyInputs,
-      boolToString(c.debugDisableExpressionsWithLazyInputs()));
+      util::boolToLowerCaseString(c.debugDisableExpressionsWithLazyInputs()));
 
   addSessionProperty(
       kDebugMemoryPoolNameRegex,
@@ -314,7 +298,7 @@ SessionProperties::SessionProperties() {
       BOOLEAN(),
       false,
       QueryConfig::kSelectiveNimbleReaderEnabled,
-      boolToString(c.selectiveNimbleReaderEnabled()));
+      util::boolToLowerCaseString(c.selectiveNimbleReaderEnabled()));
 
   addSessionProperty(
       kQueryTraceEnabled,
@@ -322,7 +306,7 @@ SessionProperties::SessionProperties() {
       BOOLEAN(),
       false,
       QueryConfig::kQueryTraceEnabled,
-      boolToString(c.queryTraceEnabled()));
+      util::boolToLowerCaseString(c.queryTraceEnabled()));
 
   addSessionProperty(
       kQueryTraceDir,
@@ -668,27 +652,8 @@ SessionProperties::SessionProperties() {
       BOOLEAN(),
       false,
       QueryConfig::kAggregationMemoryCompactionReclaimEnabled,
-      boolToString(c.aggregationMemoryCompactionReclaimEnabled()));
-}
-
-const std::string SessionProperties::toVeloxConfig(
-    const std::string& name) const {
-  auto it = sessionProperties_.find(name);
-  if (it != sessionProperties_.end() &&
-      it->second->getVeloxConfig().has_value()) {
-    return it->second->getVeloxConfig().value();
-  }
-  return name;
-}
-
-json SessionProperties::serialize() const {
-  json j = json::array();
-  json tj;
-  for (const auto& sessionProperty : sessionProperties_) {
-    protocol::to_json(tj, sessionProperty.second->getMetadata());
-    j.push_back(tj);
-  }
-  return j;
+      util::boolToLowerCaseString(
+          c.aggregationMemoryCompactionReclaimEnabled()));
 }
 
 bool SessionProperties::useVeloxGeospatialJoin() const {
