+++ presto-native-execution/presto_cpp/main/tests/CMakeLists.txt
@@ -21,7 +21,6 @@ add_executable(
   PrestoToVeloxQueryConfigTest.cpp
   QueryContextCacheTest.cpp
   ServerOperationTest.cpp
-  SessionPropertiesTest.cpp
   TaskManagerTest.cpp
   QueryContextManagerTest.cpp
   TaskInfoTest.cpp
